+++ src/main/java/ZtechAplication/repository/MarcaRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import ZtechAplication.model.Marca;
 
-public interface MarcaRepository extends CrudRepository<Marca, Integer> {
+public interface MarcaRepository extends CrudRepository<Marca, Long> {
 
 }
